feat: working workspace queries

+++ src/convex/customers.ts
@@ -1,7 +1,8 @@
-import { v } from "convex/values";
+import { ConvexError, v } from "convex/values";
 import { GenericMutationCtx } from "convex/server";
 import { mutation, query } from "./_generated/server";
 import { DataModel } from "./_generated/dataModel";
+import { getAuthData } from "./utils";
 
 type UpsertArgs = {
   name: string;
@@ -17,13 +18,17 @@ export const upsert = async (
   ctx: GenericMutationCtx<DataModel>,
   { email, name, address, city, code, country, photoUrl }: UpsertArgs,
 ) => {
+  const { workspace } = await getAuthData(ctx);
+
   const customer = await ctx.db
     .query("customers")
-    .withIndex("by_email", (q) => q.eq("email", email))
+    .withIndex("by_workspace", (q) => q.eq("workspace", workspace))
+    .filter((q) => q.eq("email", email))
     .unique();
 
   if (!customer) {
     const id = await ctx.db.insert("customers", {
+      workspace,
       name,
       email,
       address,
@@ -37,6 +42,7 @@ export const upsert = async (
   }
 
   await ctx.db.patch(customer._id, {
+    workspace,
     name: name || customer.name,
     email: email || customer.email,
     address: address || customer.address,
@@ -53,11 +59,32 @@ export const get = query({
   args: {
     id: v.id("customers"),
   },
-  handler: async (ctx, args) => await ctx.db.get(args.id),
+  handler: async (ctx, { id }) => {
+    const { workspace } = await getAuthData(ctx);
+
+    const customer = await ctx.db
+      .query("customers")
+      .withIndex("by_workspace", (q) => q.eq("workspace", workspace))
+      .filter((q) => q.eq("_id", id as string))
+      .unique();
+
+    if (!customer) {
+      throw new ConvexError("Customer not found");
+    }
+
+    return customer;
+  },
 });
 
 export const list = query({
-  handler: async (ctx) => await ctx.db.query("customers").collect(),
+  handler: async (ctx) => {
+    const { workspace } = await getAuthData(ctx);
+
+    return await ctx.db
+      .query("customers")
+      .withIndex("by_workspace", (q) => q.eq("workspace", workspace))
+      .collect();
+  },
 });
 
 export const create = mutation({
@@ -74,7 +101,10 @@ export const create = mutation({
     ctx,
     { name, email, address, city, code, country, photoUrl },
   ) => {
+    const { workspace } = await getAuthData(ctx);
+
     await ctx.db.insert("customers", {
+      workspace,
       name,
       email,
       address,
@@ -101,14 +131,27 @@ export const update = mutation({
     ctx,
     { id, name, email, address, city, code, country, photoUrl },
   ) => {
+    const { workspace } = await getAuthData(ctx);
+
+    const customer = await ctx.db
+      .query("customers")
+      .withIndex("by_workspace", (q) => q.eq("workspace", workspace))
+      .filter((q) => q.eq("_id", id as string))
+      .unique();
+
+    if (!customer) {
+      throw new ConvexError("Item not found");
+    }
+
     await ctx.db.patch(id, {
-      name,
-      email,
-      address,
-      city,
-      code,
-      country,
-      photoUrl,
+      workspace,
+      name: name || customer.name,
+      email: email || customer.email,
+      address: address || customer.address,
+      city: city || customer.city,
+      code: code || customer.code,
+      country: country || customer.country,
+      photoUrl: photoUrl || customer.photoUrl,
     });
   },
 });
@@ -118,6 +161,18 @@ export const remove = mutation({
     id: v.id("customers"),
   },
   handler: async (ctx, { id }) => {
+    const { workspace } = await getAuthData(ctx);
+
+    const customer = await ctx.db
+      .query("customers")
+      .withIndex("by_workspace", (q) => q.eq("workspace", workspace))
+      .filter((q) => q.eq("_id", id as string))
+      .unique();
+
+    if (!customer) {
+      throw new ConvexError("Customer not found in this workspace");
+    }
+
     await ctx.db.delete(id);
   },
 });
