+++ src/convex/eventTypes.ts
@@ -1,15 +1,37 @@
-import { v } from "convex/values";
+import { ConvexError, v } from "convex/values";
 import { mutation, query } from "./_generated/server";
+import { getAuthData } from "./utils";
 
 export const get = query({
   args: {
     id: v.id("eventTypes"),
   },
-  handler: async (ctx, args) => await ctx.db.get(args.id),
+  handler: async (ctx, { id }) => {
+    const { workspace } = await getAuthData(ctx);
+
+    const eventType = await ctx.db
+      .query("eventTypes")
+      .withIndex("by_workspace", (q) => q.eq("workspace", workspace))
+      .filter((q) => q.eq("_id", id as string))
+      .unique();
+
+    if (!eventType) {
+      throw new ConvexError("Item not found");
+    }
+
+    return eventType;
+  },
 });
 
 export const list = query({
-  handler: async (ctx) => await ctx.db.query("eventTypes").collect(),
+  handler: async (ctx) => {
+    const { workspace } = await getAuthData(ctx);
+
+    return await ctx.db
+      .query("eventTypes")
+      .withIndex("by_workspace", (q) => q.eq("workspace", workspace))
+      .collect();
+  },
 });
 
 export const create = mutation({
@@ -20,7 +42,15 @@ export const create = mutation({
     duration: v.number(),
   },
   handler: async (ctx, { name, variant, description, duration }) => {
-    await ctx.db.insert("eventTypes", { name, variant, description, duration });
+    const { workspace } = await getAuthData(ctx);
+
+    await ctx.db.insert("eventTypes", {
+      workspace,
+      name,
+      variant,
+      description,
+      duration,
+    });
   },
 });
 
@@ -33,7 +63,25 @@ export const update = mutation({
     duration: v.optional(v.number()),
   },
   handler: async (ctx, { id, name, variant, description, duration }) => {
-    await ctx.db.patch(id, { name, variant, description, duration });
+    const { workspace } = await getAuthData(ctx);
+
+    const eventType = await ctx.db
+      .query("eventTypes")
+      .withIndex("by_workspace", (q) => q.eq("workspace", workspace))
+      .filter((q) => q.eq("_id", id as string))
+      .unique();
+
+    if (!eventType) {
+      throw new ConvexError("Item not found");
+    }
+
+    await ctx.db.patch(id, {
+      workspace,
+      name: name || eventType.name,
+      variant: variant || eventType.variant,
+      description: description || eventType.description,
+      duration: duration || eventType.duration,
+    });
   },
 });
 
@@ -42,6 +90,18 @@ export const remove = mutation({
     id: v.id("eventTypes"),
   },
   handler: async (ctx, { id }) => {
+    const { workspace } = await getAuthData(ctx);
+
+    const eventType = await ctx.db
+      .query("eventTypes")
+      .withIndex("by_workspace", (q) => q.eq("workspace", workspace))
+      .filter((q) => q.eq("_id", id as string))
+      .unique();
+
+    if (!eventType) {
+      throw new ConvexError("EventType not found in this workspace");
+    }
+
     await ctx.db.delete(id);
   },
 });
