+++ src/convex/events.ts
@@ -1,15 +1,37 @@
-import { v } from "convex/values";
+import { ConvexError, v } from "convex/values";
 import { mutation, query } from "./_generated/server";
+import { getAuthData } from "./utils";
 
 export const get = query({
   args: {
     id: v.id("events"),
   },
-  handler: async (ctx, args) => await ctx.db.get(args.id),
+  handler: async (ctx, { id }) => {
+    const { workspace } = await getAuthData(ctx);
+
+    const event = await ctx.db
+      .query("events")
+      .withIndex("by_workspace", (q) => q.eq("workspace", workspace))
+      .filter((q) => q.eq("_id", id as string))
+      .unique();
+
+    if (!event) {
+      throw new ConvexError("Event not found");
+    }
+
+    return event;
+  },
 });
 
 export const list = query({
-  handler: async (ctx) => await ctx.db.query("events").collect(),
+  handler: async (ctx) => {
+    const { workspace } = await getAuthData(ctx);
+
+    return await ctx.db
+      .query("events")
+      .withIndex("by_workspace", (q) => q.eq("workspace", workspace))
+      .collect();
+  },
 });
 
 export const create = mutation({
@@ -22,7 +44,10 @@ export const create = mutation({
     guests: v.array(v.id("customers")),
   },
   handler: async (ctx, { end, start, title, notes, variant, guests }) => {
-    await ctx.db.insert("events", {
+    const { workspace } = await getAuthData(ctx);
+
+    const id = await ctx.db.insert("events", {
+      workspace,
       end,
       start,
       title,
@@ -30,6 +55,8 @@ export const create = mutation({
       variant,
       guests,
     });
+
+    return await ctx.db.get(id);
   },
 });
 
@@ -44,13 +71,26 @@ export const update = mutation({
     guests: v.optional(v.array(v.id("customers"))),
   },
   handler: async (ctx, { id, end, start, title, notes, variant, guests }) => {
+    const { workspace } = await getAuthData(ctx);
+
+    const event = await ctx.db
+      .query("events")
+      .withIndex("by_workspace", (q) => q.eq("workspace", workspace))
+      .filter((q) => q.eq("_id", id as string))
+      .unique();
+
+    if (!event) {
+      throw new ConvexError("Item not found");
+    }
+
     await ctx.db.patch(id, {
-      end,
-      start,
-      title,
-      notes,
-      variant,
-      guests,
+      workspace,
+      end: end || event.end,
+      start: start || event.start,
+      title: title || event.title,
+      notes: notes || event.notes,
+      variant: variant || event.variant,
+      guests: guests || event.guests,
     });
   },
 });
@@ -60,6 +100,18 @@ export const remove = mutation({
     id: v.id("events"),
   },
   handler: async (ctx, { id }) => {
+    const { workspace } = await getAuthData(ctx);
+
+    const event = await ctx.db
+      .query("events")
+      .withIndex("by_workspace", (q) => q.eq("workspace", workspace))
+      .filter((q) => q.eq("_id", id as string))
+      .unique();
+
+    if (!event) {
+      throw new ConvexError("Event not found in this workspace");
+    }
+
     await ctx.db.delete(id);
   },
 });
